Get tests going again by syncing fixtures and tests
Campaign id 42 is magic because it is created in a fixture, which avoids trying to connect to speakout. One test mocks that connection to test the code path of fetching and creating a campaign. We could maye reverse the setup to mock all the tests to use the cache, then clear the cache for that test. The advantage is we avoid having to keep the test data and the initial fixtures in sync. This came up now because of the change to using real messages for the tests.

The fixtures also needed to include more languages and countries.

The create contribution code now returns the created contributions, which the tests can use to verify the results.

-